+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 * Grouping of statements in `do`-blocks is now respected. [Issue
   74](https://github.com/tweag/ormolu/issues/74).
 
+* Comments on pragmas are now preserved. [Issue
+  216](https://github.com/tweag/ormolu/issues/216).
+
 ## Ormolu 0.0.4.0
 
 * When given several files to format, Ormolu does not stop on the first